avcodec/avcodec, avformat/movenc: support embedding channel layout to stream side data
Added support for passing complex channel layout configuration as side
packet data (AV_PKT_DATA_AUDIO_TRACK_CHANNEL_LAYOUT) to ISO media files
as specified by ISO/IEC 14496-12. AVAudioTrackChannelLayout has the
fields to setting both predefined audio layouts, completely configuring
the azimuth and elevation of each speaker as well as describing the
number of audio objects in the scene.

This information isn't integrated into the existing channel layout
system though, which is much more restricted compared to what the
standard permits. However, the side packet data is structured so that it
does not require too much ISO base media file format knowledge in client
code. In addition, it should be possible to extend the enumeration and
the record to allow for using the same side data for a more native
solution.

The names of the channels and layouts are available in
channel_layout_isoiec23001_8.h with slightly obtuse names such as
AV_SPEAKER_POSITION_ISOIEC23001_8_L and
AV_CH_LAYOUT_ISOIEC23001_8_1_0_0 to encourage path forward to a more
native solution for FFmpeg.

This channel layout information ends up to a chnl box in the written
file in an isom track.

Signed-off-by: Erkki Seppälä <erkki.seppala.ext@nokia.com>
Signed-off-by: OZOPlayer <OZOPL@nokia.com>