avcodec/metaenc, metadec: added an encoder/decoder for timed metadata
It has the codec id AV_CODEC_ID_META and type of AVMEDIA_TYPE_DATA.

This codec basically passes the data forward and is used for referring
timed meta data tracks by a codec. It is useful for dealing with the
metadata in a similar way as other kinds of codecs.

Signed-off-by: Erkki Seppälä <erkki.seppala.ext@nokia.com>
Signed-off-by: OZOPlayer <OZOPL@nokia.com>