avformat/mov: mov_read_tkhd stores alternate group to AV_PKT_DATA_TRACK_ALTERNATE_GROUP
This is done only if the option read_alternate_group is set as to
not change the behavior when copying tracks.

FFmpeg automatically assigns track groups (when no
AV_PKT_DATA_TRACK_ALTERNATE_GROUP side data is not available) by the
track indices, so if the track indices would be changed, reusing the
original alternate group from the track would likely by unwanted,
because the reassignment of alternate group would be skipped.

Signed-off-by: Erkki Seppälä <erkki.seppala.ext@nokia.com>
Signed-off-by: OZOPlayer <OZOPL@nokia.com>